refactor: Improve certificate decoding logic

+++ Sources/COpenSSL/include/pal_bio.h
@@ -1,6 +1,6 @@
 #include <openssl/bio.h>
 
-int cc_BIO_get_mem_data(BIO *bp, char **buf);
+long cc_BIO_get_mem_data(BIO *bp, char **buf);
 
 int cc_BIO_should_retry(BIO *bp);
